tweak file organization
