+++ src/book/searchPages.test.ts
@@ -1,4 +1,4 @@
-import { pageNumbersForTest, pageNumbersForSelector, formatAsRanges } from './searchPages';
+import { pageNumbersForTest, pageNumbersForSelector } from './searchPages';
 
 const pageStub = (num, str, child) => {
   const el = document.createElement('div');
@@ -29,28 +29,5 @@ describe('Finds page numbers', () => {
     ];
 
     expect(pageNumbersForSelector(pages, 'p')).toEqual([2, 3]);
-  });
-});
-
-
-test('Doesn\'t create ranges for spaced pages', () => {
-  expect(formatAsRanges([1, 3, 5, 7])).toEqual('1, 3, 5, 7');
-});
-
-describe('Make Ranges', () => {
-  test('Creates ranges for adjacent pages', () => {
-    expect(formatAsRanges([1, 2, 3, 4])).toEqual('1–4');
-  });
-
-  test('Creates multiple ranges', () => {
-    expect(formatAsRanges([1, 2, 5, 6, 7])).toEqual('1–2, 5–7');
-  });
-
-  test('Can end with a range', () => {
-    expect(formatAsRanges([1, 3, 4, 5, 8, 10, 11])).toEqual('1, 3–5, 8, 10–11');
-  });
-
-  test('Can end with a lone page', () => {
-    expect(formatAsRanges([1, 3, 4, 5, 8, 9, 12])).toEqual('1, 3–5, 8–9, 12');
   });
 });
