+++ src/utils/index.ts
@@ -1,4 +1,10 @@
 import shallowEqual from './shallowEqual';
 import { throttleFrame, throttleTime } from './throttle';
+import formatAsRanges from './formatAsRanges';
 
-export { shallowEqual, throttleFrame, throttleTime };
+export {
+    shallowEqual,
+    throttleFrame,
+    throttleTime,
+    formatAsRanges
+};
